Fix for lock nodes option bug
